Split g_measure_ks into a file in the generic_ks directory
